+++ lib/acts_as_tenant/sidekiq.rb
@@ -1,22 +1,36 @@
 module ActsAsTenant::Sidekiq
+  class BaseMiddleware
+    def self.sidekiq_7_and_up?
+      Gem::Version.new(Sidekiq::VERSION) >= Gem::Version.new("7")
+    end
+  end
+
   # Get the current tenant and store in the message to be sent to Sidekiq.
-  class Client
+  class Client < BaseMiddleware
+    include Sidekiq::ClientMiddleware if sidekiq_7_and_up?
+
     def call(worker_class, msg, queue, redis_pool)
-      msg['acts_as_tenant'] ||=
-        {
-            'class' => ActsAsTenant.current_tenant.class.name,
-            'id' => ActsAsTenant.current_tenant.id
-        } if ActsAsTenant.current_tenant.present?
+      if ActsAsTenant.current_tenant.present?
+        msg["acts_as_tenant"] ||=
+          {
+            "class" => ActsAsTenant.current_tenant.class.name,
+            "id" => ActsAsTenant.current_tenant.id
+          }
+      end
 
       yield
     end
   end
 
   # Pull the tenant out and run the current thread with it.
-  class Server
+  class Server < BaseMiddleware
+    include Sidekiq::ServerMiddleware if sidekiq_7_and_up?
+
     def call(worker_class, msg, queue)
-      if msg.has_key?('acts_as_tenant')
-        account = msg['acts_as_tenant']['class'].constantize.find msg['acts_as_tenant']['id']
+      if msg.has_key?("acts_as_tenant")
+        klass = msg["acts_as_tenant"]["class"].constantize
+        id = msg["acts_as_tenant"]["id"]
+        account = klass.class_eval(&ActsAsTenant.configuration.job_scope).find(id)
         ActsAsTenant.with_tenant account do
           yield
         end
@@ -40,6 +54,8 @@ Sidekiq.configure_server do |config|
   config.server_middleware do |chain|
     if defined?(Sidekiq::Middleware::Server::RetryJobs)
       chain.insert_before Sidekiq::Middleware::Server::RetryJobs, ActsAsTenant::Sidekiq::Server
+    elsif defined?(Sidekiq::Batch::Server)
+      chain.insert_before Sidekiq::Batch::Server, ActsAsTenant::Sidekiq::Server
     else
       chain.add ActsAsTenant::Sidekiq::Server
     end