+++ spec/dummy/app/mailers/user_mailer.rb
@@ -1,0 +1,9 @@
+class UserMailer < ApplicationMailer
+  def comment_notification
+    mail(body: "")
+  end
+
+  def welcome_email
+    mail to: params[:user].email
+  end
+end