+++ spec/dummy/app/models/aliased_task.rb
@@ -1,0 +1,4 @@
+class AliasedTask < ActiveRecord::Base
+  belongs_to :project_alias, class_name: "Project"
+  acts_as_tenant(:account)
+end