+++ spec/dummy/app/models/users_account.rb
@@ -1,0 +1,4 @@
+class UsersAccount < ActiveRecord::Base
+  belongs_to :user
+  acts_as_tenant :account
+end